The new Cell framework
In the original design, I noticed the activity related to different spaces was located in a  Space File. This file was primarily focused on creating various spaces, as outlined in the PDF requirements on Page 2. After further review, it seems feasible to merge this file with the main Space file.