The draft of creating map
The draft for creating the map involves the following steps: First, create an interface for the Map. Then, create the specific Game Legends of Valor Map. Finally, integrate the different cells from the last Cell file. The random generation process can be optimized by utilizing the Dice class for randomization.